+++ src/rebar3_grisp.erl
@@ -15,5 +15,6 @@ init(State) ->
     lists:foldl(fun(Mod, {ok, S}) -> Mod:init(S) end, {ok, State}, [
         rebar3_grisp_deploy,
         rebar3_grisp_build,
+        rebar3_grisp_package,
         rebar3_grisp_version
     ]).